reinstall lab2 and add triangle
